+++ src/adapter/adapters/cohere/adapter_impl.rs
@@ -36,7 +36,7 @@ impl Adapter for CohereAdapter {
 		INSTANCE.get_or_init(|| AdapterConfig::default().with_auth_env_name("COHERE_API_KEY"))
 	}
 
-	fn get_service_url(_kind: AdapterKind, service_type: ServiceType) -> String {
+	fn get_service_url(_model_info: ModelInfo, service_type: ServiceType) -> String {
 		match service_type {
 			ServiceType::Chat | ServiceType::ChatStream => format!("{BASE_URL}chat"),
 		}
@@ -52,11 +52,11 @@ impl Adapter for CohereAdapter {
 		let ModelInfo {
 			adapter_kind,
 			model_name,
-		} = model_info;
+		} = model_info.clone();
 
 		let stream = matches!(service_type, ServiceType::ChatStream);
 
-		let url = Self::get_service_url(adapter_kind, service_type);
+		let url = Self::get_service_url(model_info, service_type);
 
 		// -- api_key (this Adapter requires it)
 		let api_key = get_api_key_resolver(adapter_kind, config_set)?;
@@ -100,7 +100,9 @@ impl Adapter for CohereAdapter {
 		Ok(WebRequestData { url, headers, payload })
 	}
 
-	fn to_chat_response(adapter_kind: AdapterKind, web_response: WebResponse) -> Result<ChatResponse> {
+	fn to_chat_response(model_info: ModelInfo, web_response: WebResponse) -> Result<ChatResponse> {
+		let adapter_kind = model_info.adapter_kind;
+
 		let WebResponse { mut body, .. } = web_response;
 
 		// -- Get usage
@@ -120,7 +122,7 @@ impl Adapter for CohereAdapter {
 	}
 
 	fn to_chat_stream(
-		_kind: AdapterKind,
+		_model_info: ModelInfo,
 		reqwest_builder: RequestBuilder,
 		options_set: ChatRequestOptionsSet<'_, '_>,
 	) -> Result<ChatStreamResponse> {