+++ src/adapter/adapters/groq/adapter_impl.rs
@@ -31,8 +31,8 @@ impl Adapter for GroqAdapter {
 		INSTANCE.get_or_init(|| AdapterConfig::default().with_auth_env_name("GROQ_API_KEY"))
 	}
 
-	fn get_service_url(kind: AdapterKind, service_type: ServiceType) -> String {
-		OpenAIAdapter::util_get_service_url(kind, service_type, BASE_URL)
+	fn get_service_url(model_info: ModelInfo, service_type: ServiceType) -> String {
+		OpenAIAdapter::util_get_service_url(model_info, service_type, BASE_URL)
 	}
 
 	fn to_web_request_data(
@@ -45,20 +45,20 @@ impl Adapter for GroqAdapter {
 		let adapter_kind = model_info.adapter_kind;
 
 		let api_key = get_api_key_resolver(adapter_kind, config_set)?;
-		let url = Self::get_service_url(adapter_kind, service_type);
+		let url = Self::get_service_url(model_info.clone(), service_type);
 
 		OpenAIAdapter::util_to_web_request_data(model_info, url, chat_req, service_type, options_set, &api_key, false)
 	}
 
-	fn to_chat_response(kind: AdapterKind, web_response: WebResponse) -> Result<ChatResponse> {
-		OpenAIAdapter::to_chat_response(kind, web_response)
+	fn to_chat_response(model_info: ModelInfo, web_response: WebResponse) -> Result<ChatResponse> {
+		OpenAIAdapter::to_chat_response(model_info, web_response)
 	}
 
 	fn to_chat_stream(
-		kind: AdapterKind,
+		model_info: ModelInfo,
 		reqwest_builder: RequestBuilder,
 		options_set: ChatRequestOptionsSet<'_, '_>,
 	) -> Result<ChatStreamResponse> {
-		OpenAIAdapter::to_chat_stream(kind, reqwest_builder, options_set)
+		OpenAIAdapter::to_chat_stream(model_info, reqwest_builder, options_set)
 	}
 }